+++ js/actionCreators/index.ts
@@ -45,6 +45,7 @@ export {
   nextN,
   next,
   previous,
+  seekToTime,
   seekForward,
   seekBackward,
   setVolume,